Switch back to wmi, at least for now.  Neither PDH nor WMI is picking up disk changes, at least in the manner tested.
